+++ app/[locale]/projects/[projectId]/funding-search/__tests__/page.spec.tsx
@@ -35,7 +35,7 @@ const mocks = [
         funderOnly: true,
         paginationOptions: {
           type: "CURSOR",
-          limit: 50,
+          limit: 5,
         },
       },
     },
@@ -53,7 +53,7 @@ const mocks = [
             };
           }),
           totalCount: 50,
-          limit: 50,
+          limit: 5,
           nextCursor: null,
           currentOffset: null,
           hasNextPage: false,
@@ -72,7 +72,7 @@ const mocks = [
         funderOnly: true,
         paginationOptions: {
           type: "CURSOR",
-          limit: 50,
+          limit: 5,
         },
       },
     },
@@ -87,7 +87,7 @@ const mocks = [
             apiTarget: null,
           }],
           totalCount: 1,
-          limit: 50,
+          limit: 5,
           nextCursor: null,
           currentOffset: null,
           hasNextPage: false,
@@ -129,7 +129,7 @@ const mocks = [
         funderOnly: true,
         paginationOptions: {
           type: "CURSOR",
-          limit: 50,
+          limit: 5,
         },
       },
     },
@@ -159,7 +159,7 @@ const mocks = [
         funderOnly: true,
         paginationOptions: {
           type: "CURSOR",
-          limit: 50,
+          limit: 5,
         },
       },
     },
@@ -177,7 +177,7 @@ const mocks = [
             };
           }),
           totalCount: 70,
-          limit: 50,
+          limit: 5,
           nextCursor: "page2",
           currentOffset: null,
           hasNextPage: true,
@@ -198,7 +198,7 @@ const mocks = [
         paginationOptions: {
           type: "CURSOR",
           cursor: "page2",
-          limit: 50,
+          limit: 5,
         },
       },
     },
@@ -216,7 +216,7 @@ const mocks = [
             };
           }),
           totalCount: 70,
-          limit: 50,
+          limit: 5,
           nextCursor: null,
           currentOffset: null,
           hasNextPage: true,