+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ### Added
+- Added new `Start DMP` page at `projects/[projectId]/dmp/start` to direct user to create new plan or upload existing [#956]
 - Added `autosave` back to the `PlanOverviewQuestionpage` [#944]
 - Hooked up the `Download plan` page and added a `download-narrative` api endpoint [#313]
 
 ### Updated
+- Improved project creation steps by updating pages in the flow (header, home, plan dashboard, create project details,and funding search) to help users in creating a plan [#956]
 - Updated Project Details subdomains field to only display once a user selects a research domain [#947]
 - Updated `Remove` buttons to be `secondary` buttons, rather than `red` [#964]